Parse data from txt file
