+++ accounts/src/main/scala/com/tms/app/AccountsController.scala
@@ -3,7 +3,9 @@ package com.tms.app
 import org.scalatra._
 import org.json4s.{DefaultFormats, Formats}
 import org.scalatra.json._
+
 import com.tms.db.Persistence
+import com.tms.models.{Account, Error}
 
 class AccountsController extends ScalatraServlet with JacksonJsonSupport {
 
@@ -19,9 +21,11 @@ class AccountsController extends ScalatraServlet with JacksonJsonSupport {
   }
 
   get("/accounts") {
-    Persistence.getAll()
+    try {
+      Persistence.getAll()
+    } catch {
+      case e: Throwable => Error(500, e.getMessage)
+    }
   }
 
 }
-
-case class Account(email: String, name: String, password: String)
