+++ accounts/src/main/scala/com/tms/models/Account.scala
@@ -1,0 +1,3 @@
+package com.tms.models
+
+case class Account(email: String, name: String, password: String)
