+++ accounts/src/main/scala/com/tms/models/Error.scala
@@ -1,0 +1,3 @@
+package com.tms.models
+
+case class Error(status: Int, message: String)
